Dashboard components extracted and fixed
