+++ app/src/utils/sortByProperty.ts
@@ -1,0 +1,29 @@
+import { SortOrder } from "../types/typeInterfaces";
+
+export default function sortByProperty<T>(
+  array: T[],
+  property: keyof T,
+  order: SortOrder = "asc",
+): T[] {
+  if (property === "name") {
+    return array.slice().sort((a, b) => {
+      if (a[property] < b[property]) {
+        return order === "desc" ? -1 : 1;
+      } else if (a[property] > b[property]) {
+        return order === "desc" ? 1 : -1;
+      } else {
+        return 0;
+      }
+    });
+  } else {
+    return array.slice().sort((a, b) => {
+      if (a[property] < b[property]) {
+        return order === "asc" ? -1 : 1;
+      } else if (a[property] > b[property]) {
+        return order === "asc" ? 1 : -1;
+      } else {
+        return 0;
+      }
+    });
+  }
+}
